address comments and move encoding functions into one util file
