+++ cstrml/swork/src/lib.rs
@@ -31,6 +31,9 @@ use sp_std::collections::btree_map::BTreeMap;
 /// Provides crypto and other std functions by implementing `runtime_interface`
 pub mod api;
 
+/// Provides util functions
+pub mod utils;
+
 #[cfg(test)]
 mod mock;
 
@@ -750,12 +753,11 @@ impl<T: Trait> Module<T> {
         sig: &SworkerSignature
     ) -> bool {
         // 1. Encode
-        let block_number_bytes = Self::encode_u64_to_string_vec_u8(block_number);
-        let reserved_bytes = Self::encode_u64_to_string_vec_u8(reserved);
-        let used_bytes = Self::encode_u64_to_string_vec_u8(used);
-        let added_files_bytes = Self::encode_files(added_files);
-        let deleted_files_bytes = Self::encode_files(deleted_files);
-
+        let block_number_bytes = utils::encode_u64_to_string_to_vec_u8(block_number);
+        let reserved_bytes = utils::encode_u64_to_string_to_vec_u8(reserved);
+        let used_bytes = utils::encode_u64_to_string_to_vec_u8(used);
+        let added_files_bytes = utils::encode_files(added_files);
+        let deleted_files_bytes = utils::encode_files(deleted_files);
 
         // 2. Construct work report data
         //{
@@ -784,88 +786,6 @@ impl<T: Trait> Module<T> {
         ].concat();
 
         api::crypto::verify_p256_sig(curr_pk, &data, sig)
-        // api::crypto::verify_work_report_sig(
-        //     curr_pk,
-        //     prev_pk,
-        //     block_number,
-        //     block_hash,
-        //     reserved,
-        //     used,
-        //     srd_root,
-        //     files_root,
-        //     added_files,
-        //     deleted_files,
-        //     sig
-        // )
-    }
-
-    // 127 -> "127" -> 49 50 55
-    fn encode_u64_to_string_vec_u8(number: u64) -> Vec<u8> {
-        let mut value = number;
-        let mut encoded_number: Vec<u8> = [].to_vec();
-        loop {
-            encoded_number.push((value%10) as u8 + 48u8); // "0" is 48u8
-            value /= 10;
-            if value == 0 {
-                break;
-            }
-        }
-        encoded_number.reverse();
-        encoded_number
-    }
-
-    fn encode_file_root(fs: Vec<u8>) -> Vec<u8> {
-        let mut rst: Vec<u8> = [].to_vec();
-        for v in fs.iter() {
-            rst.extend(Self::encode_u8_to_hex_string_to_vec_u8(*v));
-        }
-        rst
-    }
-
-    // 91 -> 5b -> "5b" -> 53 98
-    fn encode_u8_to_hex_string_to_vec_u8(number: u8) -> Vec<u8> {
-        let upper_value = number / 16 as u8;
-        let lower_value = number % 16 as u8;
-        [Self::encode_u8_to_hex_char_to_u8(upper_value), Self::encode_u8_to_hex_char_to_u8(lower_value)].to_vec()
-    }
-
-    // 5 -> "5" -> 53
-    // 11 -> "b" -> 98
-    fn encode_u8_to_hex_char_to_u8(number: u8) -> u8 {
-        if number < 10u8 {
-            return number + 48u8;
-        } else {
-            return number - 10u8 + 97u8;
-        }
-    }
-
-    fn encode_files(fs: &Vec<(Vec<u8>, u64)>) -> Vec<u8> {
-        // "["
-        let open_square_brackets_bytes: Vec<u8> = [91].to_vec();
-        // "\"hash\":\""
-        let hash_bytes: Vec<u8> = [123, 34, 104, 97, 115,104, 34, 58, 34].to_vec();
-        // "\",\"size\":"
-        let size_bytes: Vec<u8> = [34, 44, 34, 115, 105, 122, 101, 34, 58].to_vec();
-        // "}"
-        let close_curly_brackets_bytes: Vec<u8> = [125].to_vec();
-        // ","
-        let comma_bytes: Vec<u8> = [44].to_vec();
-        // "]"
-        let close_square_brackets_bytes: Vec<u8> = [93].to_vec();
-        let mut rst: Vec<u8> = open_square_brackets_bytes.clone();
-        let len = fs.len();
-        for (pos, (hash, size)) in fs.iter().enumerate() {
-            rst.extend(hash_bytes.clone());
-            rst.extend(Self::encode_file_root(hash.clone()));
-            rst.extend(size_bytes.clone());
-            rst.extend(Self::encode_u64_to_string_vec_u8(*size));
-            rst.extend(close_curly_brackets_bytes.clone());
-            if pos != len-1 { rst.extend(comma_bytes.clone()) }
-        }
-
-        rst.extend(close_square_brackets_bytes.clone());
-
-        rst
     }
 
     fn get_current_block_number() -> BlockNumber {
